style: dev.handsup.common.entity -> dev.handsup.common.domain 네이밍 변경 (#79)
