+++ core/src/main/java/dev/handsup/auth/domain/Auth.java
@@ -2,7 +2,6 @@ package dev.handsup.auth.domain;
 
 import static jakarta.persistence.GenerationType.*;
 
-import dev.handsup.common.entity.TimeBaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -11,33 +10,35 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import dev.handsup.common.domain.TimeBaseEntity;
+
 @Entity
 @Getter
 @NoArgsConstructor
 public class Auth extends TimeBaseEntity {
 
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-	@Column(name = "auth_id")
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    @Column(name = "auth_id")
+    private Long id;
 
-	@Column(name = "auth_user_id", nullable = false, unique = true)
-	private Long userId;
+    @Column(name = "auth_user_id", nullable = false, unique = true)
+    private Long userId;
 
-	@Column(name = "auth_refresh_token", nullable = false, unique = true)
-	private String refreshToken;
+    @Column(name = "auth_refresh_token", nullable = false, unique = true)
+    private String refreshToken;
 
-	@Builder
-	private Auth(Long userId, String refreshToken) {
-		this.userId = userId;
-		this.refreshToken = refreshToken;
-	}
+    @Builder
+    private Auth(Long userId, String refreshToken) {
+        this.userId = userId;
+        this.refreshToken = refreshToken;
+    }
 
-	public static Auth of(Long userId, String refreshToken) {
-		return Auth.builder()
-			.userId(userId)
-			.refreshToken(refreshToken)
-			.build();
-	}
+    public static Auth of(Long userId, String refreshToken) {
+        return Auth.builder()
+            .userId(userId)
+            .refreshToken(refreshToken)
+            .build();
+    }
 
 }
