+++ core/src/main/java/dev/handsup/auth/domain/BlacklistToken.java
@@ -2,7 +2,6 @@ package dev.handsup.auth.domain;
 
 import static jakarta.persistence.GenerationType.*;
 
-import dev.handsup.common.entity.TimeBaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -12,28 +11,30 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import dev.handsup.common.domain.TimeBaseEntity;
+
 @Entity
 @Getter
 @NoArgsConstructor
 @Table(name = "blacklist_token")
 public class BlacklistToken extends TimeBaseEntity {
 
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-	@Column(name = "blacklist_token_id")
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    @Column(name = "blacklist_token_id")
+    private Long id;
 
-	@Column(name = "refresh_token", nullable = false, unique = true)
-	private String refreshToken;
+    @Column(name = "refresh_token", nullable = false, unique = true)
+    private String refreshToken;
 
-	@Builder
-	private BlacklistToken(String refreshToken) {
-		this.refreshToken = refreshToken;
-	}
+    @Builder
+    private BlacklistToken(String refreshToken) {
+        this.refreshToken = refreshToken;
+    }
 
-	public static BlacklistToken from(String refreshToken) {
-		return BlacklistToken.builder()
-			.refreshToken(refreshToken)
-			.build();
-	}
+    public static BlacklistToken from(String refreshToken) {
+        return BlacklistToken.builder()
+            .refreshToken(refreshToken)
+            .build();
+    }
 }
