+++ core/src/main/java/dev/handsup/bidding/domain/Bidding.java
@@ -6,11 +6,6 @@ import static jakarta.persistence.FetchType.*;
 import static jakarta.persistence.GenerationType.*;
 import static lombok.AccessLevel.*;
 
-import dev.handsup.auction.domain.Auction;
-import dev.handsup.bidding.exception.BiddingErrorCode;
-import dev.handsup.common.entity.TimeBaseEntity;
-import dev.handsup.common.exception.ValidationException;
-import dev.handsup.user.domain.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Enumerated;
@@ -23,97 +18,103 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import dev.handsup.auction.domain.Auction;
+import dev.handsup.bidding.exception.BiddingErrorCode;
+import dev.handsup.common.domain.TimeBaseEntity;
+import dev.handsup.common.exception.ValidationException;
+import dev.handsup.user.domain.User;
+
 @Entity
 @Getter
 @NoArgsConstructor(access = PROTECTED)
 public class Bidding extends TimeBaseEntity {
 
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-	@Column(name = "bidding_id")
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    @Column(name = "bidding_id")
+    private Long id;
 
-	@Column(name = "bidding_price", nullable = false)
-	private int biddingPrice;
+    @Column(name = "bidding_price", nullable = false)
+    private int biddingPrice;
 
-	@ManyToOne(fetch = LAZY)
-	@JoinColumn(name = "auction_id",
-		nullable = false,
-		foreignKey = @ForeignKey(NO_CONSTRAINT))
-	private Auction auction;
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "auction_id",
+        nullable = false,
+        foreignKey = @ForeignKey(NO_CONSTRAINT))
+    private Auction auction;
 
-	@ManyToOne(fetch = LAZY)
-	@JoinColumn(name = "bidder_id",
-		nullable = false,
-		foreignKey = @ForeignKey(NO_CONSTRAINT))
-	private User bidder;
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "bidder_id",
+        nullable = false,
+        foreignKey = @ForeignKey(NO_CONSTRAINT))
+    private User bidder;
 
-	@Enumerated(STRING)
-	@Column(name = "trading_status", nullable = false)
-	private TradingStatus tradingStatus;
+    @Enumerated(STRING)
+    @Column(name = "trading_status", nullable = false)
+    private TradingStatus tradingStatus;
 
-	@Builder
-	private Bidding(int biddingPrice, Auction auction, User bidder) {
-		auction.validateNotSeller(bidder);
-		this.biddingPrice = biddingPrice;
-		this.auction = auction;
-		this.bidder = bidder;
-		this.tradingStatus = TradingStatus.WAITING;
-	}
+    @Builder
+    private Bidding(int biddingPrice, Auction auction, User bidder) {
+        auction.validateNotSeller(bidder);
+        this.biddingPrice = biddingPrice;
+        this.auction = auction;
+        this.bidder = bidder;
+        this.tradingStatus = TradingStatus.WAITING;
+    }
 
-	// 테스트용
-	public Bidding(Long id, int biddingPrice, Auction auction, User bidder, TradingStatus tradingStatus) {
-		auction.validateNotSeller(bidder);
-		this.id = id;
-		this.biddingPrice = biddingPrice;
-		this.auction = auction;
-		this.bidder = bidder;
-		this.tradingStatus = tradingStatus;
-	}
+    // 테스트용
+    public Bidding(Long id, int biddingPrice, Auction auction, User bidder, TradingStatus tradingStatus) {
+        auction.validateNotSeller(bidder);
+        this.id = id;
+        this.biddingPrice = biddingPrice;
+        this.auction = auction;
+        this.bidder = bidder;
+        this.tradingStatus = tradingStatus;
+    }
 
-	// 테스트용
-	public Bidding(int biddingPrice, Auction auction, User bidder, TradingStatus tradingStatus) {
-		auction.validateNotSeller(bidder);
-		this.biddingPrice = biddingPrice;
-		this.auction = auction;
-		this.bidder = bidder;
-		this.tradingStatus = tradingStatus;
-	}
+    // 테스트용
+    public Bidding(int biddingPrice, Auction auction, User bidder, TradingStatus tradingStatus) {
+        auction.validateNotSeller(bidder);
+        this.biddingPrice = biddingPrice;
+        this.auction = auction;
+        this.bidder = bidder;
+        this.tradingStatus = tradingStatus;
+    }
 
-	public static Bidding of(int biddingPrice, Auction auction, User bidder) {
-		return Bidding.builder()
-			.biddingPrice(biddingPrice)
-			.auction(auction)
-			.bidder(bidder)
-			.build();
-	}
+    public static Bidding of(int biddingPrice, Auction auction, User bidder) {
+        return Bidding.builder()
+            .biddingPrice(biddingPrice)
+            .auction(auction)
+            .bidder(bidder)
+            .build();
+    }
 
-	// 비즈니스 메서드
-	public void updateTradingStatusComplete() {
-		if (tradingStatus != TradingStatus.PROGRESSING) {
-			throw new ValidationException(BiddingErrorCode.CAN_NOT_COMPLETE_TRADING);
-		}
-		tradingStatus = TradingStatus.COMPLETED;
-	}
+    // 비즈니스 메서드
+    public void updateTradingStatusComplete() {
+        if (tradingStatus != TradingStatus.PROGRESSING) {
+            throw new ValidationException(BiddingErrorCode.CAN_NOT_COMPLETE_TRADING);
+        }
+        tradingStatus = TradingStatus.COMPLETED;
+    }
 
-	public void updateTradingStatusCanceled() {
-		if (tradingStatus != TradingStatus.PROGRESSING) {
-			throw new ValidationException(BiddingErrorCode.CAN_NOT_CANCEL_TRADING);
-		}
-		tradingStatus = TradingStatus.CANCELED;
-	}
+    public void updateTradingStatusCanceled() {
+        if (tradingStatus != TradingStatus.PROGRESSING) {
+            throw new ValidationException(BiddingErrorCode.CAN_NOT_CANCEL_TRADING);
+        }
+        tradingStatus = TradingStatus.CANCELED;
+    }
 
-	public void updateTradingStatusPreparing() {
-		if (tradingStatus != TradingStatus.WAITING) {
-			throw new ValidationException(BiddingErrorCode.CAN_NOT_PREPARE_TRADING);
-		}
-		tradingStatus = TradingStatus.PREPARING;
-	}
+    public void updateTradingStatusPreparing() {
+        if (tradingStatus != TradingStatus.WAITING) {
+            throw new ValidationException(BiddingErrorCode.CAN_NOT_PREPARE_TRADING);
+        }
+        tradingStatus = TradingStatus.PREPARING;
+    }
 
-	public void updateTradingStatusProgressing() {
-		if (tradingStatus == TradingStatus.COMPLETED) {
-			throw new ValidationException(BiddingErrorCode.CAN_NOT_PROGRESS_TRADING);
-		}
-		tradingStatus = TradingStatus.PROGRESSING;
-	}
+    public void updateTradingStatusProgressing() {
+        if (tradingStatus == TradingStatus.COMPLETED) {
+            throw new ValidationException(BiddingErrorCode.CAN_NOT_PROGRESS_TRADING);
+        }
+        tradingStatus = TradingStatus.PROGRESSING;
+    }
 }
