+++ core/src/main/java/dev/handsup/comment/domain/Comment.java
@@ -5,9 +5,6 @@ import static jakarta.persistence.FetchType.*;
 import static jakarta.persistence.GenerationType.*;
 import static lombok.AccessLevel.*;
 
-import dev.handsup.auction.domain.Auction;
-import dev.handsup.common.entity.TimeBaseEntity;
-import dev.handsup.user.domain.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.ForeignKey;
@@ -19,58 +16,62 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import dev.handsup.auction.domain.Auction;
+import dev.handsup.common.domain.TimeBaseEntity;
+import dev.handsup.user.domain.User;
+
 @Entity
 @Getter
 @NoArgsConstructor(access = PROTECTED)
 public class Comment extends TimeBaseEntity {
 
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-	@Column(name = "comment_id")
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    @Column(name = "comment_id")
+    private Long id;
 
-	@Column(name = "content", nullable = false)
-	private String content;
+    @Column(name = "content", nullable = false)
+    private String content;
 
-	@ManyToOne(fetch = LAZY)
-	@JoinColumn(name = "auction_id",
-		nullable = false,
-		foreignKey = @ForeignKey(NO_CONSTRAINT))
-	private Auction auction;
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "auction_id",
+        nullable = false,
+        foreignKey = @ForeignKey(NO_CONSTRAINT))
+    private Auction auction;
 
-	@ManyToOne(fetch = LAZY)
-	@JoinColumn(name = "writer_id",
-		nullable = false,
-		foreignKey = @ForeignKey(NO_CONSTRAINT))
-	private User writer;
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "writer_id",
+        nullable = false,
+        foreignKey = @ForeignKey(NO_CONSTRAINT))
+    private User writer;
 
-	@Builder
-	private Comment(String content, Auction auction, User writer) {
-		this.content = content;
-		this.auction = auction;
-		this.writer = writer;
-	}
+    @Builder
+    private Comment(String content, Auction auction, User writer) {
+        this.content = content;
+        this.auction = auction;
+        this.writer = writer;
+    }
 
-	// 테스트용 생성자
-	private Comment(Long id, String content, Auction auction, User writer) {
-		this.id = id;
-		this.content = content;
-		this.auction = auction;
-		this.writer = writer;
-	}
+    // 테스트용 생성자
+    private Comment(Long id, String content, Auction auction, User writer) {
+        this.id = id;
+        this.content = content;
+        this.auction = auction;
+        this.writer = writer;
+    }
 
-	public static Comment of(String content, Auction auction, User writer) {
-		return Comment.builder()
-			.content(content)
-			.auction(auction)
-			.writer(writer)
-			.build();
-	}
+    public static Comment of(String content, Auction auction, User writer) {
+        return Comment.builder()
+            .content(content)
+            .auction(auction)
+            .writer(writer)
+            .build();
+    }
 
-	public static Comment getTestComment(
-		Long id, String content, Auction auction, User writer
-	) {
-		return new Comment(id, content, auction, writer);
-	}
+    public static Comment getTestComment(
+        Long id, String content, Auction auction, User writer
+    ) {
+        return new Comment(id, content, auction, writer);
+    }
 
 }
