+++ core/src/main/java/dev/handsup/notification/domain/Notification.java
@@ -6,8 +6,6 @@ import static jakarta.persistence.FetchType.*;
 import static jakarta.persistence.GenerationType.*;
 import static lombok.AccessLevel.*;
 
-import dev.handsup.auction.domain.Auction;
-import dev.handsup.common.entity.TimeBaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Enumerated;
@@ -20,64 +18,67 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import dev.handsup.auction.domain.Auction;
+import dev.handsup.common.domain.TimeBaseEntity;
+
 @Entity
 @Getter
 @NoArgsConstructor(access = PROTECTED)
 public class Notification extends TimeBaseEntity {
 
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-	@Column(name = "notification_id")
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    @Column(name = "notification_id")
+    private Long id;
 
-	@Column(name = "sender_email", nullable = false)
-	private String senderEmail;
+    @Column(name = "sender_email", nullable = false)
+    private String senderEmail;
 
-	@Column(name = "receiver_email", nullable = false)
-	private String receiverEmail;
+    @Column(name = "receiver_email", nullable = false)
+    private String receiverEmail;
 
-	@Column(name = "content", nullable = false)
-	private String content;
+    @Column(name = "content", nullable = false)
+    private String content;
 
-	@Column(name = "type", nullable = false)
-	@Enumerated(STRING)
-	private NotificationType type;
+    @Column(name = "type", nullable = false)
+    @Enumerated(STRING)
+    private NotificationType type;
 
-	@ManyToOne(fetch = LAZY)
-	@JoinColumn(name = "auction_id",
-		nullable = false,
-		foreignKey = @ForeignKey(NO_CONSTRAINT))
-	private Auction auction;
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "auction_id",
+        nullable = false,
+        foreignKey = @ForeignKey(NO_CONSTRAINT))
+    private Auction auction;
 
-	@Builder
-	private Notification(
-		String senderEmail,
-		String receiverEmail,
-		String content,
-		NotificationType type,
-		Auction auction
-	) {
-		this.senderEmail = senderEmail;
-		this.receiverEmail = receiverEmail;
-		this.content = content;
-		this.type = type;
-		this.auction = auction;
-	}
+    @Builder
+    private Notification(
+        String senderEmail,
+        String receiverEmail,
+        String content,
+        NotificationType type,
+        Auction auction
+    ) {
+        this.senderEmail = senderEmail;
+        this.receiverEmail = receiverEmail;
+        this.content = content;
+        this.type = type;
+        this.auction = auction;
+    }
 
-	public static Notification of(
-		String senderEmail,
-		String receiverEmail,
-		String content,
-		NotificationType type,
-		Auction auction
-	) {
-		return Notification.builder()
-			.senderEmail(senderEmail)
-			.receiverEmail(receiverEmail)
-			.content(content)
-			.type(type)
-			.auction(auction)
-			.build();
-	}
+    public static Notification of(
+        String senderEmail,
+        String receiverEmail,
+        String content,
+        NotificationType type,
+        Auction auction
+    ) {
+        return Notification.builder()
+            .senderEmail(senderEmail)
+            .receiverEmail(receiverEmail)
+            .content(content)
+            .type(type)
+            .auction(auction)
+            .build();
+    }
 
 }
